% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/AnnealingFunctions4SpatialResponseSurfaceSampling.R
\name{anneal}
\alias{anneal}
\title{Simulated Annealing}
\usage{
anneal(
  mysample,
  candidates,
  dpnt,
  weight = 0,
  phi = NULL,
  T_ini = 1,
  coolingRate = 0.95,
  maxPermuted = 20 * nrow(mysample),
  maxNoChange = 20,
  verbose = getOption("verbose")
)
}
\arguments{
\item{mysample}{\code{\link{data.frame}} of units of initial sample,
with point_id, two spatial coordinates, two principal component scores,
and associated design-point.}

\item{candidates}{\code{\link{data.frame}} with candidate units per design
point. \code{\link{data.frame}} has same variables as \code{mysample}.}

\item{dpnt}{\code{\link{data.frame}} with coordinates of design points of a
rotatable central composite design for two factors.}

\item{weight}{weight for the average distance of the sampling points to the}

\item{phi}{distance parameter of exponential semivariogram (optional);
if \code{NULL} (default value) the geometric criterion is used,
else the model-based criterion is used.}

\item{T_ini}{initial temperature of annealing schedule.}

\item{coolingRate}{cooling rate of annealing schedule.}

\item{maxPermuted}{maximum number of permutations per chain.}

\item{maxNoChange}{stopping criterion. Maximum number of successive chains
without change of the criterion.}

\item{verbose}{\code{\link{logical}} for printing the temperature, the
criterion, the number of accepted proposals per chain,
and the number of improvements per chain during the optimisation.}
}
\value{
\code{\link{list}} with \code{\link{data.frame}} of optimised sample,
\code{\link{numeric}} with minimised criterion, and \code{\link{numeric}}
with trace of criterion.
}
\description{
This function optimises the sampling pattern of a
central composite response surface design sample.
}
