% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectiveFunctions4MBSamplingVariogram.R
\name{FIM}
\alias{FIM}
\title{Fisher Information Matrix}
\usage{
FIM(A, D, model, thetas, perturbation)
}
\arguments{
\item{A}{correlation matrix of sampling points.}

\item{D}{distance matrix of sampling points.}

\item{model}{\pkg{gstat} model type of a priori semivariogram model.}

\item{thetas}{parameters of semivariogram model.}

\item{perturbation}{proportion of a semivariogram parameter value added to
that value.}
}
\value{
Fisher information matrix.
}
\description{
This function computes the Fisher information matrix.
}
\seealso{
\code{\link[gstat]{vgm}}
}
