% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grdAmazonia}
\alias{grdAmazonia}
\title{Aboveground Live Woody Biomass (Eastern Amazonia)}
\format{
A \code{\link[tibble]{tibble}} with 1306296 rows and 10 columns:
\describe{
\item{x1}{Easting (m).}
\item{x2}{Northing (m).}
\item{AGB}{aboveground live woody biomass (megatons per ha).}
\item{SWIR2}{long term mean of MODIS short-wave infrared radiation.}
\item{Terra_PP}{primary production (kg C/m2).}
\item{Prec_dm}{average precipitation in driest month (mm).}
\item{Elevation}{elevation (m).}
\item{Clay}{clay content (g/kg).}
\item{Biome}{four biomes.}
\item{Ecoregion}{sixteen ecoregions.}
}
}
\usage{
grdAmazonia
}
\description{
Aboveground live woody biomass (AGB) in megatons per ha
(Baccini et al., 2012) in a rectangular area of 1,642 km x 928 km in Eastern
Amazonia (Brazil). The data of Baccini et al. (2012) were aggregated to a map
with a resolution of 1 km x 1 km.
}
\details{
Map projection: World Sinusoidal (esri:54008).
}
\references{
Baccini, A., S. J. Goetz, W. S. Walker, N. T. Laporte, M. Sun,
D. Sulla-Menashe, J. Hackler, P. S. A. Beck, R. Dubayah, M.A. Friedl,
S. Samanta and R. A. Houghton (2012). Estimated carbon dioxide emissions
from tropical deforestation improved by carbon-density maps.
Nature Climate Change, 2(3):182–185.
}
\keyword{datasets}
