% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grdAmhara}
\alias{grdAmhara}
\title{Covariates West Amhara (Ethiopa)}
\format{
A \code{\link[tibble]{tibble}} with 10843 rows and 7 columns:
\describe{
\item{s1}{Easting (m).}
\item{s2}{Northing (m).}
\item{dem}{elevation (m).}
\item{evi}{enhanced vegetation index.}
\item{rfl_NIR}{near-infrared reflectance.}
\item{rfl_red}{red reflectance.}
\item{lst}{land surface temperature.}
}
}
\usage{
grdAmhara
}
\description{
Covariates at the nodes of a 1 km x 1 km grid in the western part of the
Amhara district in Ethiopia.
}
\details{
Map projection: WGS84 / LAEA (laea +ellps=WGS84 +lat_0=5 +lon_0=20).
}
\seealso{
\link{sampleAmhara}
}
\keyword{datasets}
