% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grdHunterValley}
\alias{grdHunterValley}
\title{Terrain Attributes for Hunter Valley (Australia)}
\format{
A \code{\link[tibble]{tibble}} with 22124 rows and 7 columns:

\describe{
\item{s1}{Easting (m).}
\item{s2}{Northing (m).}
\item{elevation_m}{elevation (m).}
\item{slope_deg}{slope (degrees).}
\item{cos_aspect}{cosine of aspect.}
\item{cti}{compound topographic index.}
\item{ndvi}{normalised difference vegetation index.}
}

The resolution of the grid is 25 m x 25 m.
}
\usage{
grdHunterValley
}
\description{
A \code{\link[tibble]{tibble}} of five terrain attributes for Hunter Valley.
}
\details{
Map projection: WGS84 / UTM zone 56S (epsg:32756).
}
\keyword{datasets}
