% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grdIberia}
\alias{grdIberia}
\title{Annual Mean Air Temperature in Spain and Portugal}
\format{
A \code{\link[tibble]{tibble}} with 957198 rows and 6 columns:
\describe{
\item{x}{Easting (m).}
\item{y}{Northing (m).}
\item{TAS2004}{Annual mean air temperature in 2004 (degrees Celsius).}
\item{TAS2009}{Annual mean air temperature in 2009 (degrees Celsius).}
\item{TAS2014}{Annual mean air temperature in 2014 (degrees Celsius).}
\item{TAS2019}{Annual mean air temperature in 2019 (degrees Celsius).}
}
}
\usage{
grdIberia
}
\description{
Annual mean air temperature at two metres above the earth surface (TAS) in
degrees C, in Spain and Portugal (islands excluded) for the years 2004, 2009,
2014, and 2019. These data are part of the CHELSA data set
(Karger et al. 2017). The resolution of the grid is about 780 m x 780 m.
}
\details{
Map projection: ETRS89-extended / LAEA Europe (epsg:3035).
}
\references{
Karger, D. N. et al. 2017. Climatologies at high resolution for
the earth’s land surface areas. Sci. Data 4:170122
doi: 10.1038/sdata.2017.122.
}
\keyword{datasets}
