% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grdKandahar}
\alias{grdKandahar}
\title{Poppy Area in the Province of Kandahar (Afghanistan)}
\format{
A \code{\link[tibble]{tibble}} with 965 rows and 4 columns:
\describe{
\item{s1}{Easting (m).}
\item{s2}{Northing (m).}
\item{poppy}{poppy area (ha) per 5 km square.}
\item{agri}{agricultural area (ha) per 5 km square.}
}
}
\usage{
grdKandahar
}
\description{
Simulated data of poppy area (ha) per 5 km square in the province of Kandahar
(Afghanistan).
}
\details{
Map projection: WGS84 / UTM zone 41N (epsg:32641).
}
\keyword{datasets}
