% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grdVoorst}
\alias{grdVoorst}
\title{Simulated Soil Organic Matter Concentration in Voorst (The Netherlands)}
\format{
A \code{\link[tibble]{tibble}} with 7528 rows and 4 columns:
\describe{
\item{s1}{Easting (m)}
\item{s2}{Northing (m)}
\item{z}{Simulated Soil Organic Matter Concentration (g/kg)}
\item{stratum}{five combinations of soil type and land use classes.
The first letter stands for the soil type: B for beekeerdgrond (sandy
wetland soil with gleyic properties), E for enkeerdgrond (sandy soil with
thick anthropogenic humic topsoil), P for podzols (sandy soil with
eluviated horizon below the topsoil), R for river clay soil,
and X for sandy soils. The second letter is for land use: A for agriculture
(grassland, arable land), and F for forest.}
}
}
\usage{
grdVoorst
}
\description{
Simulated data on a 25 m x 25 grid of the organic matter concentration
(g/kg) in the topsoil of an 6 km x 1 km area in the Netherlands.
}
\details{
Map projection: Amersfoort / RDnew (epsg:28992).
}
\keyword{datasets}
