% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grdXuancheng}
\alias{grdXuancheng}
\title{Terrain Attributes and Climate Variables in Xuancheng (China)}
\format{
A \code{\link[tibble]{tibble}} with 113847 rows and 11 columns:
\describe{
\item{s1}{Easting (m).}
\item{s2}{Northing (m).}
\item{plan.curvature}{planar curvature.}
\item{profile.curvature}{profile curvature.}
\item{slope}{slope (degrees).}
\item{temperature}{temperature.}
\item{precipitation}{precipitation.}
\item{twi}{topographic wetness index.}
\item{dem}{elevation (m).}
\item{SOM_RF}{predictions of the organic matter concentration (g/kg) in
the A horizon of the soil, obtained by random forest.}
\item{SOM_KED}{predictions of the organic matter concentration (g/kg) in
the A horizon of the soil, obtained by kriging with an external drift}
}
}
\usage{
grdXuancheng
}
\description{
Five terrain attributes and two climate variables in Xuancheng (Anhui
province, China) and predictions of soil organic matter concentration
obtained by random forest (RF) and kriging with an external drift (KED).
}
\details{
Map projection: WGS84 / UTM zone 50N (epsg:32560).
}
\seealso{
\link{sampleXuancheng}
}
\keyword{datasets}
