% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectiveFunctions4MBSamplingVariogram.R
\name{logdet}
\alias{logdet}
\title{Determinant Information Matrix}
\usage{
logdet(points, model, thetas, perturbation = 0.01)
}
\arguments{
\item{points}{\code{\link{data.frame}} or \code{SpatialPoints(DataFrame)}
with coordinates of sampling points.}

\item{model}{\pkg{gstat} model type of a priori semivariogram model.}

\item{thetas}{parameters of semivariogram model.}

\item{perturbation}{proportion of a semivariogram parameter value added to
that value.}
}
\value{
log of the reciprocal determinant of Fisher information matrix.
}
\description{
This function computes the log of the reciprocal determinant of Fisher
information matrix.
}
\seealso{
\code{\link[gstat]{vgm}}
}
