% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectiveFunctions4MBSamplingVariogram.R
\name{MAKV}
\alias{MAKV}
\title{Mean Augmented Kriging Variance}
\usage{
MAKV(points, esample, model, thetas, perturbation = 0.01)
}
\arguments{
\item{points}{\code{\link{data.frame}} with coordinates of fixed and free sampling points}

\item{esample}{\code{\link[sp]{SpatialPoints}} used for evaluation.}

\item{model}{\pkg{gstat} model type of a priori semivariogram model.}

\item{thetas}{parameters of semivariogram model.}

\item{perturbation}{proportion of a semivariogram parameter value added to
that value.}
}
\value{
mean augmented kriging variance.
}
\description{
Mean Augmented Kriging Variance
}
