% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/AnnealingFunctions4SpatialResponseSurfaceSampling.R
\name{permute}
\alias{permute}
\title{Permutation}
\usage{
permute(mysample, candidates)
}
\arguments{
\item{mysample}{\code{\link{data.frame}} of units of current sample.}

\item{candidates}{\code{\link{data.frame}} with candidate units per design.
point. This \code{\link{data.frame}} has same variables as \code{mysample}.}
}
\value{
\code{\link{data.frame}} with units of permuted sample.
}
\description{
This function permutes a current sample by replacing one randomly selected
unit by a unit randomly selected from the remaining candidate units of the
same design point.
}
