% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleAmhara}
\alias{sampleAmhara}
\title{Soil Organic Matter West Amhara (Ethiopa)}
\format{
A \code{\link[tibble]{tibble}} with 407 rows and 14 columns:
\describe{
\item{s1}{Easting (m).}
\item{s2}{Northing (m).}
\item{SOM}{organic matter concentration in dag/kg in the topsoil.}
\item{soil}{soil type.}
\item{dem}{elevation (m).}
\item{twi}{topographic wetness index.}
\item{slope}{slope.}
\item{evi}{enhanced vegetation index.}
\item{aspect}{aspect.}
\item{rfl_MIR}{ mid-infrared reflectance.}
\item{rfl_NIR}{near-infrared reflectance.}
\item{rfl_red}{red reflectance.}
\item{lst}{land surface temperature.}
\item{lst_night}{land surface temperature in the night.}
}
}
\usage{
sampleAmhara
}
\description{
Convenience sample of measurements of organic matter concentration in dag/kg
in the topsoil at 407 points (SOM), mainly along roads, in West Amhara
(Ethiopia). Besides, several auxiliary variables are in the \code{\link[tibble]{tibble}}.
}
\details{
Map projection: WGS84 / LAEA (laea +ellps=WGS84 +lat_0=5 +lon_0=20).
}
\seealso{
\link{grdAmhara}
}
\keyword{datasets}
