% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleCRF}
\alias{sampleCRF}
\title{Soil Electrical Conductivity and Electromagnetic Induction (Uzbekistan)}
\format{
A \code{\link[tibble]{tibble}} with 142 rows and 6 columns:
\describe{
\item{x}{Easting (m).}
\item{y}{Northing (m).}
\item{ECe150}{electrical conductivity 150 cm.}
\item{EMv1m}{electromagnetic induction (vertical dipole mode, 1m).}
\item{date}{date.}
\item{survey}{survey identifier.}
}
}
\usage{
sampleCRF
}
\description{
Data of the electrical conductivity until a depth of 150 cm (ECe150) and of
the electromagnetic induction in vertical dipole mode with transmitter at
1 m (EMv1m) at 142 points in the Cotton Research Farm in Uzbekistan.
The data have been collected in eight surveys in the period 2008-2011. The
date is the column date, and the period is the column survey.
}
\details{
Map projection:  WGS84 / UTM zone 41N (epsg:32641).
}
\seealso{
\link{grdCRF}
}
\keyword{datasets}
