% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleMelle}
\alias{sampleMelle}
\title{Nitrate-N data for Melle (Belgium)}
\format{
A \code{\link[tibble]{tibble}} with 30 rows and 3 columns:
\describe{
\item{s1}{Easting (m).}
\item{s2}{Northing (m).}
\item{N}{Nitrate-N (NO3-N) in kg/ha in the soil layer 0-90 cm.}
}
}
\usage{
sampleMelle
}
\description{
Nitrate-N (NO3-N) in kg/ha in the soil layer 0-90 cm, using a standard soil
density of 1,500 kg/m3, at 30 points in an agricultural field in Melle
(Belgium).
}
\details{
Map projection: ETRS89 / Belgian Lambert 2008 (EPSG:3812).
}
\keyword{datasets}
