% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleXuancheng}
\alias{sampleXuancheng}
\title{Organic Matter Concentration by Three Sampling Designs in Xuancheng (China)}
\format{
A \code{\link[tibble]{tibble}} with 183 rows and 12 columns:
\describe{
\item{s1}{Easting (m).}
\item{s2}{Northing (m).}
\item{SOM_A_hori}{organic matter concentration (g/kg) in the A horizon.}
\item{sample}{sampling design: grid, iPSM, or STSI.}
\item{stratum}{strata used in STSI}
\item{plan.curvature}{planar curvature.}
\item{profile.curvature}{profile curvature.}
\item{slope}{slope (degrees).}
\item{temperature}{temperature.}
\item{precipitation}{precipitation.}
\item{twi}{topographic wetness index.}
\item{dem}{elevation (m).}
}
}
\usage{
sampleXuancheng
}
\description{
Data of the organic matter concentration (g/kg) in the A horizon
of the soil at 183 points in Xuancheng
(Anhui province, China). The data are collected by three sampling designs,
indicated in the column \code{sample} of the \code{\link[tibble]{tibble}}: a square grid (grid),
sampling tailored to mapping with the individual predictive soil mapping
method (iPSM), and stratified simple random sampling (STSI).
}
\details{
Map projection: WGS84 / UTM zone 50N (epsg:32560).
}
\seealso{
\link{grdXuancheng}
}
\keyword{datasets}
