% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObjectiveFunctions4MBSamplingVariogram.R
\name{validate}
\alias{validate}
\title{Validate input files}
\usage{
validate(sample, model, thetas, perturbation)
}
\arguments{
\item{sample}{\code{\link[sp]{SpatialPoints}} used for prediction at the
evaluation points or used for evaluation.}

\item{model}{\pkg{gstat} model type of a priori semivariogram model.}

\item{thetas}{parameters of semivariogram model.}

\item{perturbation}{proportion of a semivariogram parameter value added to
that value.}
}
\description{
This function checks various input files
}
