% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/AnnealingFunctions4SpatialResponseSurfaceSampling.R
\name{validate_colnames}
\alias{validate_colnames}
\title{Validate Column Names}
\usage{
validate_colnames(x)
}
\arguments{
\item{x}{data.frame to validate}
}
\value{
validated data frame
}
\description{
This function checks whether all columns of data frame with sampling points
and of data frame with candidate sampling points are available
}
